+++ Modules/Intent.Modules.Constants/Designers.cs
@@ -1,0 +1,7 @@
+﻿namespace Intent.Modules.Constants;
+
+public static class Designers
+{
+    public const string Domain = "Domain";
+    public const string Services = "Services";
+}