+++ Modules/Intent.Modules.Security.Shared/SecurityType.cs
@@ -1,0 +1,8 @@
+
+namespace Intent.Modules.Security.Shared;
+
+public enum SecurityType
+{
+    Role,
+    Policy
+}