+++ Modules/Intent.Modules.AspNetCore.Mvc/release-notes.md
@@ -1,4 +1,5 @@
 ### Version 1.0.0
 
+- Improvement: `Roles` and `Policies` will now used constants instead of strings, if they are available.
 - Improvement: Included module help topic.
 Initial release.