+++ Modules/Intent.Modules.AspNetCore/release-notes.md
@@ -1,10 +1,11 @@
 ### Version 6.0.8
 
-- Fixed: When using minimal hosting model some service registration statements would be generated after application configuration statements.
+- Improvement: `Roles` and `Policies` now output as constants in a `Permissions` file.
 
 ### Version 6.0.7
 
 - Fixed: `Configuration` not resolving correctly for Application Builder registrations
+- Fixed: When using minimal hosting model some service registration statements would be generated after application configuration statements.
 
 ### Version 6.0.6