+++ Modules/Intent.Modules.EntityFrameworkCore.DataMasking/release-notes.md
@@ -1,3 +1,7 @@
+### Version 1.0.1
+
+- Improvement: `Roles` and `Policies` will now used constants instead of strings, if they are available.
+
 ### Version 1.0.0
 
 - New: Initial module